+++ pkg/util/tests/kubernetes.go
@@ -314,6 +314,12 @@ func CreateObjects(t *testing.T, k8s kubernetes.Interface, arango arangoClientSe
 			vl := *v
 			_, err := arango.PlatformV1alpha1().ArangoPlatformCharts(vl.GetNamespace()).Create(context.Background(), vl, meta.CreateOptions{})
 			require.NoError(t, err)
+		case **platformApi.ArangoPlatformService:
+			require.NotNil(t, v)
+
+			vl := *v
+			_, err := arango.PlatformV1alpha1().ArangoPlatformServices(vl.GetNamespace()).Create(context.Background(), vl, meta.CreateOptions{})
+			require.NoError(t, err)
 		default:
 			require.Fail(t, fmt.Sprintf("Unable to create object: %s", reflect.TypeOf(v).String()))
 		}
@@ -533,6 +539,12 @@ func UpdateObjects(t *testing.T, k8s kubernetes.Interface, arango arangoClientSe
 			vl := *v
 			_, err := arango.PlatformV1alpha1().ArangoPlatformCharts(vl.GetNamespace()).Update(context.Background(), vl, meta.UpdateOptions{})
 			require.NoError(t, err)
+		case **platformApi.ArangoPlatformService:
+			require.NotNil(t, v)
+
+			vl := *v
+			_, err := arango.PlatformV1alpha1().ArangoPlatformServices(vl.GetNamespace()).Update(context.Background(), vl, meta.UpdateOptions{})
+			require.NoError(t, err)
 		default:
 			require.Fail(t, fmt.Sprintf("Unable to create object: %s", reflect.TypeOf(v).String()))
 		}
@@ -716,6 +728,11 @@ func DeleteObjects(t *testing.T, k8s kubernetes.Interface, arango arangoClientSe
 
 			vl := *v
 			require.NoError(t, arango.PlatformV1alpha1().ArangoPlatformCharts(vl.GetNamespace()).Delete(context.Background(), vl.GetName(), meta.DeleteOptions{}))
+		case **platformApi.ArangoPlatformService:
+			require.NotNil(t, v)
+
+			vl := *v
+			require.NoError(t, arango.PlatformV1alpha1().ArangoPlatformServices(vl.GetNamespace()).Delete(context.Background(), vl.GetName(), meta.DeleteOptions{}))
 		default:
 			require.Fail(t, fmt.Sprintf("Unable to delete object: %s", reflect.TypeOf(v).String()))
 		}
@@ -1237,6 +1254,21 @@ func RefreshObjects(t *testing.T, k8s kubernetes.Interface, arango arangoClientS
 			} else {
 				*v = vn
 			}
+		case **platformApi.ArangoPlatformService:
+			require.NotNil(t, v)
+
+			vl := *v
+
+			vn, err := arango.PlatformV1alpha1().ArangoPlatformServices(vl.GetNamespace()).Get(context.Background(), vl.GetName(), meta.GetOptions{})
+			if err != nil {
+				if kerrors.IsNotFound(err) {
+					*v = nil
+				} else {
+					require.NoError(t, err)
+				}
+			} else {
+				*v = vn
+			}
 		default:
 			require.Fail(t, fmt.Sprintf("Unable to get object: %s", reflect.TypeOf(v).String()))
 		}
@@ -1495,6 +1527,14 @@ func SetMetaBasedOnType(t *testing.T, object meta.Object) {
 		v.SetSelfLink(fmt.Sprintf("/api/%s/%s/%s/%s",
 			platformApi.SchemeGroupVersion.String(),
 			platform.ArangoPlatformChartResourcePlural,
+			object.GetNamespace(),
+			object.GetName()))
+	case *platformApi.ArangoPlatformService:
+		v.Kind = platform.ArangoPlatformServiceResourceKind
+		v.APIVersion = networkingApi.SchemeGroupVersion.String()
+		v.SetSelfLink(fmt.Sprintf("/api/%s/%s/%s/%s",
+			platformApi.SchemeGroupVersion.String(),
+			platform.ArangoPlatformServiceResourcePlural,
 			object.GetNamespace(),
 			object.GetName()))
 	default: