+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - (Feature) (Platform) MetaV1 Integration Service
 - (Feature) (Platform) Chart Overrides
 - (Feature) Parallel Executor
+- (Feature) (Platform) Service Handler
 
 ## [1.2.49](https://github.com/arangodb/kube-arangodb/tree/1.2.49) (2025-06-17)
 - (Maintenance) Optimize go.mod